메모 작성시 빈줄이 구현되지 않았던 오류 수정 (titleText와 contentText 분리하는 방법 개선)

+++ Memo/Scene/Write/WriteViewController.swift
@@ -49,7 +49,6 @@ class WriteViewController: BaseViewController {
     override func configure() {
 
         mainView.userTextView.text = (task?.title ?? "") + "\n" + (task?.content ?? "")
-        
         mainView.userTextView.becomeFirstResponder()
         
         let sharedButton = UIBarButtonItem(image: UIImage(systemName: "square.and.arrow.up"), style: .plain, target: self, action: #selector(sharedButtonClicked))
@@ -70,18 +69,14 @@ class WriteViewController: BaseViewController {
     }
     
     func updateMemo() {
-        contentArray = mainView.userTextView.text.split(separator: "\n").map{String($0)}
+        guard let memoText = mainView.userTextView.text else { return }
         
+        contentArray = memoText.split(separator: "\n").map{String($0)}
         
         if contentArray.count > 1 {
             titleText = contentArray[0]
-            contentText = ""
-            for i in 1...contentArray.count - 1 {
-                contentText! += contentArray[i]
-                if i != contentArray.count - 1 {
-                    contentText! += "\n"
-                }
-            }
+            let startIndex = memoText.index(memoText.startIndex, offsetBy: titleText.count + 1)
+            contentText = String(memoText[startIndex...])
         } else if contentArray.count == 0 {
             if let task = task {
                 repository.deleteTask(task: task)
